move OpProtoAndCheckerMaker from operator to op_proto_maker
